+++ internal/dao/mongo/key_dao.go
@@ -5,7 +5,7 @@ import (
     "errors"
     "slices"
 
-    "github.com/i2-open/i2goSignals/internal/dao/ids"
+    "github.com/i2-open/i2goSignals/pkg/dao/ids"
     "github.com/i2-open/i2goSignals/internal/dao/interfaces"
     "github.com/i2-open/i2goSignals/pkg/logger"
     "go.mongodb.org/mongo-driver/v2/bson"
