+++ internal/providers/dbProviders/mongo_provider/test/rebind_test.go
@@ -7,7 +7,7 @@ import (
     "testing"
     "time"
 
-    "github.com/i2-open/i2goSignals/internal/dao/ids"
+    "github.com/i2-open/i2goSignals/pkg/dao/ids"
     "github.com/i2-open/i2goSignals/internal/dao/interfaces"
     "github.com/i2-open/i2goSignals/internal/providers/dbProviders/mongo_provider"
     "github.com/stretchr/testify/assert"
